rustcodepro
a complete python package for the Toxannotator. Tomorrow final release.

It implements the Keras and tensorflow for the LSTM and Convolutional Neural Network for the classification with my optimized settings for the sample post translational modification data.

I code only RUST but 1 hour daily i write python and daily 1 hour i read python and hence it was so easy to code this in python.

Started this evening an finished it. Tomorrow final release.